Remove unecessary duplicate merge for Chi Burst.
I noticed that brewmaster had the same id twice for "Healing" and "Damage". Investigated which ID was used for the Damage and noticed it was the same as in other specs, so the "Chi Burst" in all specs has the same functionaly, and the other specs does not have their own entry for the talent.