Add alias for the special essence font that is triggered by Faeline Stomp.
When using Faeline Stomp as mistweaver, there's a bonus feature, "Up to 5 allies caught in the faline are healed with an Essence Font bolt", and this seems to use a different ID than the normal Essence Font. Could maybe aruge it should be grouped by itself, a target can have both hots at the same time.